+++ src/main/java/com/driveu/server/domain/file/dao/FileRepository.java
@@ -1,4 +1,4 @@
-package com.driveu.server.domain.resource.dao;
+package com.driveu.server.domain.file.dao;
 
 import com.driveu.server.domain.resource.domain.File;
 import org.springframework.data.jpa.repository.JpaRepository;
