+++ src/main/java/com/driveu/server/domain/note/dto/response/NoteUpdateTitleResponse.java
@@ -1,4 +1,4 @@
-package com.driveu.server.domain.resource.dto.response;
+package com.driveu.server.domain.note.dto.response;
 
 import com.driveu.server.domain.resource.domain.Note;
 import lombok.AllArgsConstructor;
