+++ src/main/java/com/driveu/server/domain/resource/application/ResourceService.java
@@ -15,6 +15,12 @@ import com.driveu.server.domain.resource.dto.response.ResourceDeleteResponse;
 import com.driveu.server.domain.resource.dto.response.ResourceFavoriteResponse;
 import com.driveu.server.domain.resource.dto.response.ResourceResponse;
 import com.driveu.server.domain.resource.dto.response.TagResponse;
+import com.driveu.server.domain.file.dao.FileRepository;
+import com.driveu.server.domain.resource.domain.File;
+import com.driveu.server.domain.link.dao.LinkRepository;
+import com.driveu.server.domain.resource.domain.Link;
+import com.driveu.server.domain.note.dao.NoteRepository;
+import com.driveu.server.domain.resource.domain.Note;
 import com.driveu.server.domain.user.dao.UserRepository;
 import com.driveu.server.domain.user.domain.User;
 import jakarta.persistence.EntityNotFoundException;
